Completely new approach to relocations to try to fit all relocation types for our platforms of interest all under one abstraction:
Relocation calculation functions now return an AST of an expression that calculates the effect of the relocation.  Interpreting this calculation is external to the interpretation of the relocation type, and is done on a per-ABI basis.  All relocation types fit into this new abstraction barring COPY types, which require a special case (as they are a special case)...